Simplify implementation
