--- src/JsonBundle/Resources/config/services.yml
@@ -1,5 +1,0 @@
-services:
-    json.listener:
-        class: AFieldGuideToElephpants\JsonBundle\EventListener\Json
-        tags:
-            - { name: kernel.event_listener, event: sculpin.core.after_format }
